Implement ability to get units for a property
This feature is important to allow a property owner to get the units for a specific property. If the property is not found, the request to the endpoint will return a 404. We are currently validating to see if the list of UnitDTOs is empty. We are using JPQL queries, this is for performance reasons especially when there will be more features and relationships with new models and Property

issue #11